Rename context module to newcontext to resolve namespace conflict
